% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workaround.R
\name{AdjustLLSTitles}
\alias{AdjustLLSTitles}
\title{Adjust the Titles of the LLS Limesurvey file}
\usage{
AdjustLLSTitles(file)
}
\arguments{
\item{file}{The LLS file}
}
\description{
This helper function adjusts the titles of the LLS Limesurvey file.
 It extracts the title from the LLS File, matches it with Meta Data and
 exports the new LLS file. By doing so, this helper function automates the
 process of adjusting the titles of the LLS file. This function is a workaround
 until we get the code, variable and plot name via the API. Once we get the
 code, variable and plot name via the API, this function will be depreciated.
}
