% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMetaData.R
\name{buildMetaMaster}
\alias{buildMetaMaster}
\title{Build Meta Master Data}
\usage{
buildMetaMaster(path, export = FALSE)
}
\arguments{
\item{path}{The path to the Excel file.}

\item{export}{If TRUE, the MetaMaster will be exported as an Excel file.}
}
\description{
This function builds the MetaMaster from a given Excel file.
}
