% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMetaData.R
\name{buildOverallReport}
\alias{buildOverallReport}
\title{Build Overall Report based on Package Name}
\usage{
buildOverallReport(packagename)
}
\arguments{
\item{packagename}{The path to the Excel file.}
}
\description{
This function will build an overall report based on the package name.
}
