% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_SurveyChecks.R
\name{check_distinct}
\alias{check_distinct}
\title{Check if All Variables of the Meta Data are Distinct}
\usage{
check_distinct(ubb)
}
\description{
This function checks if all variables names of the meta data
are distinct. It counts if a variable name is used more than once and returns
a data frame with the results.
}
\examples{
\dontrun{
check_distinct(ubb = TRUE)
}
}
