% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_SurveyChecks.R
\name{check_ExtraPlot}
\alias{check_ExtraPlot}
\title{Check if All Extra Variables are Ploted Distinctly}
\usage{
check_ExtraPlot()
}
\description{
There are several extra plots, which means that some variables
 of LimeSurvey appear in more than one plot. This function checks how often
   the variable appears in a report and if distinct number of plots are used.
}
