% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_manualmeta.R
\name{check_manualmeta}
\alias{check_manualmeta}
\title{Check the manual data}
\usage{
check_manualmeta()
}
\arguments{
\item{path}{Path to the Excel file with metadata}
}
\value{
A message from the DB_send function indicating if the table was
 successfully uploaded to the database
}
\description{
This function reads an Excel file with metadata and uploads
 each sheet as a table in the database.
}
