% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_SurveyChecks.R
\name{check_MasterTemplates}
\alias{check_MasterTemplates}
\title{Get the templates}
\usage{
check_MasterTemplates()
}
\description{
Some text in LimeSurvey are stored in HTML format.
 This helper function extracts it from the HTML code.
}
\examples{
check_MasterTemplates()
}
