% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_SurveyChecks.R
\name{check_SurveyTemplates}
\alias{check_SurveyTemplates}
\title{Check Survey Templates}
\usage{
check_SurveyTemplates()
}
\description{
Some text in LimeSurvey are stored in HTML format.
 This helper function extracts it from the HTML code.
}
