% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_fun.R
\name{connect_DB}
\alias{connect_DB}
\title{Connect to the Database}
\usage{
DB_send(table = readxl::read_excel("myfile.xlsx"), name = "myname")
}
\value{
A message indicating if the table was successfully uploaded.
}
\description{
This function connects to the database using the credentials
 provided in the config file.
}
