% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_config}
\alias{create_config}
\title{Create a Config File}
\usage{
create_config()
}
\arguments{
\item{file}{The name of the YAML file to create. Default is "config.yml".}

\item{export}{If TRUE, the function writes the YAML file. Otherwise, it returns the YAML.}
}
\description{
This function creates a YAML config file with default values.
}
