% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_TestSchools}
\alias{create_TestSchools}
\title{Create Test Schools}
\usage{
create_TestSchools()
}
\description{
This function creates a data frame to test all master templates.
 It creates example surveys for each master template to upload in a test environment.
}
