% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_LSS.R
\name{DB_DeleteFrom}
\alias{DB_DeleteFrom}
\title{Delete Observations From a Table}
\usage{
DB_send(table = readxl::read_excel("myfile.xlsx"), name = "myname")
}
\arguments{
\item{table}{Data frame or table that will be uploaded.}
}
\value{
A message indicating if the table was successfully deleted.
}
\description{
This function deletes all observations from a table in a PostgreSQL database.
}
