% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_fun.R
\name{DB_MetaUpdate}
\alias{DB_MetaUpdate}
\title{Upload Metadata to the Database}
\usage{
DB_MetaUpdate(path)
}
\arguments{
\item{path}{Path to the Excel file with metadata}
}
\value{
A message from the DB_send function indicating if the table was
 successfully uploaded to the database
}
\description{
This function reads an Excel file with metadata and uploads
 each sheet as a table in the database.
}
\examples{
\dontrun{
DB_MetaUpdate(path = "data/report_meta_dev.xlsx")
}
}
