% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_fun.R
\name{DB_send}
\alias{DB_send}
\title{Send a Table to the Database}
\usage{
DB_send(table, name)
}
\arguments{
\item{table}{Data frame to be uploaded to the database}

\item{name}{Name of the table in the database}
}
\value{
A message indicating if the table was successfully uploaded to the database
}
\description{
This function send a table to the database.
}
\examples{
\dontrun{
  DB_send(table = readxl::read_excel("myfile.xlsx"), name = "master_to_template")
}
}
