% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_LSS.R
\name{DB_UploadLSS}
\alias{DB_UploadLSS}
\title{Upload All LSS Survey Files to the Database}
\usage{
DB_send(table = readxl::read_excel("myfile.xlsx"), name = "myname")
}
\arguments{
\item{dir}{Directory where the .lss files are stored.}
}
\value{
A message indicating if the .lss files were successfully uploaded.
}
\description{
This function uploads all .lss files in a directory to a PostgreSQL database.
}
