% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_XML}
\alias{export_XML}
\title{Export XML/LSS Files from a Data Frame}
\usage{
export_XML(data)
}
\arguments{
\item{data}{Data}
}
\description{
This function exports XML files from a data frame to the "lssfiles" directory.
}
\examples{
\dontrun{
export_XML(lss_surveys)
}
}
