% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GetQuestions.R
\name{get_AllMasterMeta}
\alias{get_AllMasterMeta}
\title{Adjust: Get ALL meta data from LimeSurvey master templates.}
\usage{
get_AllMasterMeta(export = FALSE)
}
\arguments{
\item{export}{Export the data to an Excel file.}
}
\value{
Results from the API.
}
\description{
This function gets the meta data from LimeSurvey master templates.
}
\examples{
\dontrun{
get_AllMasterMeta(export = FALSE)
}
}
