% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{get_master}
\alias{get_master}
\title{Deprecated: Get the templates}
\usage{
get_master(templatename)
}
\arguments{
\item{templatename}{Template name}
}
\description{
Some text in LimeSurvey are stored in HTML format.
 This helper function extracts it from the HTML code.
}
