% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GetQuestions.R
\name{get_MasterMeta}
\alias{get_MasterMeta}
\title{Get meta data from LimeSurvey master templates.}
\usage{
get_MasterMeta(id, name)
}
\arguments{
\item{id}{Survey ID.}

\item{name}{Survey name (surveyls_title).}
}
\value{
Results from the API.
}
\description{
This function gets the meta data from LimeSurvey master templates.
}
\examples{
\dontrun{
get_MasterMeta()
}
}
