% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{get_masters}
\alias{get_masters}
\title{Deprecated: Get the templates Deprecated}
\usage{
get_masters()
}
\arguments{
\item{templatename}{Template name}
}
\description{
Some text in LimeSurvey are stored in HTML format.
This helper function extracts it from the HTML code.

Some text in LimeSurvey are stored in HTML format.
This helper function extracts it from the HTML code.

Some text in LimeSurvey are stored in HTML format.
This helper function extracts it from the HTML code.
}
