% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GetQuestions.R
\name{get_MasterTemplate}
\alias{get_MasterTemplate}
\title{Get master templates from LimeSurvey}
\usage{
get_MasterTemplate(sart = NULL)
}
\arguments{
\item{sart}{School type.}
}
\value{
Results from the API.
}
\description{
This function gets the master templates from Lime Survey.
}
\examples{
\dontrun{
get_MasterTemplate(sart = "gs")
}
}
