% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metalist}
\alias{get_metalist}
\title{Get individual surveys from Overallreports}
\usage{
get_metalist()
}
\arguments{
\item{overalls}{List of indivudal surveys}

\item{data}{Meta data
Get a list with unique meta data}
}
\description{
This helper function extracts the individual surveys (parents, kids, etc.)
for the overall reports.

Append overall reports

This helper function appends the overall reports for one
survey package

This function creates a list with unique meta data based on the
input from the Master to template file.
}
