% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{get_TemplateDF}
\alias{get_TemplateDF}
\title{Get the templates}
\usage{
get_TemplateDF(mastername)
}
\arguments{
\item{mastername}{Master name}
}
\description{
Some text in LimeSurvey are stored in HTML format.
This helper function extracts it from the HTML code.
}
