% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{get_templates}
\alias{get_templates}
\title{Get the templates Deprecated}
\usage{
get_templates()
}
\description{
Some text in LimeSurvey are stored in HTML format.
 This helper function extracts it from the HTML code.
}
