% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{gisela_report}
\alias{gisela_report}
\title{Get the templates}
\usage{
gisela_report(template)
}
\arguments{
\item{template}{Template name}
}
\description{
Some text in LimeSurvey are stored in HTML format.
 This helper function extracts it from the HTML code.
}
