% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{joinMetaGisela}
\alias{joinMetaGisela}
\title{Get the templates}
\usage{
joinMetaGisela(templatename, mastername, mistakes = TRUE, update = FALSE)
}
\arguments{
\item{templatename}{Template name}

\item{mastername}{Master name}

\item{mistakes}{Logical}
}
\description{
Some text in LimeSurvey are stored in HTML format.
This helper function extracts it from the HTML code.
}
