% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_delete.R
\name{Limer_DeleteSurvey}
\alias{Limer_DeleteSurvey}
\title{Delete a Survey from Lime Survey Instance}
\usage{
Limer_DeleteSurvey(id)
}
\arguments{
\item{id}{The id of the survey to delete.}
}
\value{
Response from the API.
}
\description{
This function deletes a survey from a Lime Survey instance.
}
