% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GET.R
\name{Limer_GetMasterTemplates}
\alias{Limer_GetMasterTemplates}
\title{Get master templates from LimeSurvey}
\usage{
Limer_GetMasterTemplates(template = FALSE)
}
\arguments{
\item{template}{Logical. If TRUE, the function will return the template name.}
}
\value{
Results from the API.
}
\description{
This function gets the master templates from Lime Survey.
}
\examples{
\dontrun{
get_MasterTemplate()
}
}
