% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_Question_properties.R
\name{Limer_GetQlist}
\alias{Limer_GetQlist}
\title{Get the list of questions for a given survey ID}
\usage{
Limer_GetQlist(id)
}
\arguments{
\item{id}{Survey ID}
}
\value{
A character vector of unique question IDs
}
\description{
This function retrieves the list of questions for a given survey ID.
}
