% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_Question_properties.R
\name{Limer_getQuestionsbyQID}
\alias{Limer_getQuestionsbyQID}
\title{Get Question Title, Plot Name, and Question Text by Question ID}
\usage{
Limer_getQuestionsbyQID(qid)
}
\arguments{
\item{qid}{Question ID}
}
\value{
A tibble with columns 'title', 'plot', and 'question'
}
\description{
This function retrieves the question title, plot name,
and question text for a given question ID.
}
