% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_Post.R
\name{Limer_sendSurveys}
\alias{Limer_sendSurveys}
\title{Send a list of survey templates to Lime Survey}
\usage{
Limer_sendSurveys()
}
\value{
Results from the API.
}
\description{
This function sends a survey template to the LimeSurvey API.
}
