% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer.R
\name{LS_Connect}
\alias{LS_Connect}
\title{Connect with Limesurvey}
\usage{
LS_Connect(server, user, credential)
}
\arguments{
\item{server}{Sever name}

\item{user}{Limesurvey user}

\item{credential}{Limesurvey credential}
}
\value{
session id aus Limesurvey
}
\description{
Verbindung mit Limesurvey aufbauen
}
