% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_delete.R
\name{LS_DeleteMasterTemplates}
\alias{LS_DeleteMasterTemplates}
\title{Delete All Mastertemplates from LimeSurvey}
\usage{
LS_DeleteMasterTemplates()
}
\value{
Response from the API.
}
\description{
This function deletes all master templates from Lime Survey that are listed
 in the master_to_template table.
}
\examples{
\dontrun{
LS_DeleteMasterTemplates()
}
}
