% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GET.R
\name{LS_GetMasterData}
\alias{LS_GetMasterData}
\title{Get the Master Data from Lime Survey}
\usage{
LS_GetMasterData(export = FALSE)
}
\arguments{
\item{export}{Export the data to an Excel file.}
}
\value{
Results from the API.
}
\description{
This function gets the raw meta data from Lime Survey for all master templates.
}
\examples{
\dontrun{
LS_GetMasterData(export = FALSE)
}

}
