% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GET.R
\name{LS_GetMasterTemplates}
\alias{LS_GetMasterTemplates}
\title{Get Master Templates from Lime Survey}
\usage{
LS_GetMasterTemplates(template = FALSE)
}
\arguments{
\item{template}{If TRUE, the function will return the template name as well.}
}
\value{
Results from the API.
}
\description{
This function gets the master templates from Lime Survey.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
LS_GetMasterTemplates(template = FALSE)
\dontshow{\}) # examplesIf}
}
