% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_Question_properties.R
\name{LS_GetQlist}
\alias{LS_GetQlist}
\title{Superseded: Get the list of questions for a given survey ID}
\usage{
LS_GetQlist(id = "197865")
}
\arguments{
\item{id}{Survey ID}
}
\value{
A character vector of unique question IDs
}
\description{
This function retrieves the list of questions for a given survey ID.
}
