% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer.R
\name{LS_Release}
\alias{LS_Release}
\title{Release a LimeSurvey API session key}
\usage{
LS_Release()
}
\description{
This function clears the LimeSurvey API session key currently in use, effectively logging out.
}
\examples{
\dontrun{
release_session_key()
}
}
