% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_Post.R
\name{LS_SendSurvey}
\alias{LS_SendSurvey}
\title{Send a survey template to Lime Survey}
\usage{
LS_SendSurvey(lss, name = NULL)
}
\arguments{
\item{lss}{The survey template.}

\item{name}{The optional name of the survey.}
}
\value{
Results from the API.
}
\description{
This function sends a survey template to the LimeSurvey.
}
\examples{
\dontrun{
LS_SendSurvey(lss = "struktur_LimeSurvey.lss",
name = "Edgar")
}
}
