% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_Post.R
\name{LS_SendSurveys}
\alias{LS_SendSurveys}
\title{Send a list of survey templates to Lime Survey}
\usage{
LS_SendSurveys()
}
\value{
Results from the API.
}
\description{
This function sends a survey template to the LimeSurvey API.
}
