% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GET.R
\name{LS_UploadTemplates}
\alias{LS_UploadTemplates}
\title{Upload Master Templates as Survey Templates to Lime Survey}
\usage{
LS_UploadTemplates()
}
\value{
Results from the API.
}
\description{
This function uploads all master templates as survey templates
 to Lime Survey.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
LS_UploadTemplates()
\dontshow{\}) # examplesIf}
}
