% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{prepare_OverallReport}
\alias{prepare_OverallReport}
\title{Build Overall Report based on Package Name}
\usage{
prepare_OverallReport(packagename)
}
\arguments{
\item{packagename}{The path to the Excel file.}
}
\description{
This function will build an overall report based on the package name.
}
