% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{prepare_RawMeta}
\alias{prepare_RawMeta}
\title{Prepare the Raw Meta Data}
\usage{
prepare_RawMeta(path, export = FALSE)
}
\arguments{
\item{path}{The path to the Excel file.}

\item{export}{If TRUE, the MetaMaster will be exported as an Excel file.}
}
\description{
This function builds the MetaMaster from a given Excel file.
}
