% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_SurveyItems}
\alias{prepare_SurveyItems}
\title{Prepare Survey Items}
\usage{
prepare_SurveyItems(string)
}
\arguments{
\item{string}{Text string}
}
\description{
This function prepares survey items by normalizing spaces, removing parentheses,
 handling special cases, and replacing certain strings.
}
\examples{
prepare_SurveyItems(string = "Die Betreuerinnen und Betreuer können sich gut ...")
}
