% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_and_combine}
\alias{remove_and_combine}
\title{Remove White Spaces and Combine Text}
\usage{
remove_and_combine(input_vector)
}
\arguments{
\item{input_vector}{A character vector.}
}
\description{
Some master templates have white spaces between words. This
 function removes the white spaces and combines the text again.
}
\examples{
remove_and_combine(input_vector = "What   a difference  a day makes.")
}
