% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{send_Report}
\alias{send_Report}
\title{Send Test Results}
\usage{
send_Report()
}
\description{
This function send the test results to the specified email address.
}
\examples{
\dontrun{
send_testrun(sendto = 'john.doe@johndoe.com')
}
}
