% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMasterReport.R
\name{send_testrun}
\alias{send_testrun}
\title{Send Test Results}
\usage{
send_testrun(sendto)
}
\arguments{
\item{sendto}{The recipient of the email}
}
\description{
This function send the test results to the specified email address.
}
\examples{
\dontrun{
send_testrun(sendto = 'john.doe@johndoe.com')
}
}
