% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{testrun}
\alias{testrun}
\title{Get the templates}
\usage{
testrun(export = FALSE)
}
\arguments{
\item{export}{Export results to Excel}
}
\description{
Some text in LimeSurvey are stored in HTML format.
 This helper function extracts it from the HTML code.
}
