% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_allMastersLimeSurvey}
\alias{update_allMastersLimeSurvey}
\title{Update all Masters LimeSurvey}
\usage{
update_allMastersLimeSurvey()
}
\description{
Some text in LimeSurvey are stored in HTML format.
This helper function extracts it from the HTML code.
}
