% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workaround.R
\name{workaround}
\alias{workaround}
\title{Workaround (Soft depreciated)}
\usage{
workaround(sid, master_title)
}
\arguments{
\item{sid}{The survey id}

\item{master_title}{The title of the master template}
}
\description{
This helper function matches the manual meta data with master
 templates from LimeSurvey. This way, the function check which variables
 of the manual meta data can be merged. Furthermore, the function adds the
 plot name to the variable name. Once we get code, variable and names and
 plot via the API, this function will be depreciated. Thus, this is a helper
 function to create a running system without any manual steps.
}
