% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_fun.R
\name{connect_DB}
\alias{connect_DB}
\title{Connect to the Database}
\usage{
connect_DB()
}
\value{
Return the connection object.
}
\description{
This function connects to the database using the credentials
 provided in the config file.
}
