% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_fun.R
\name{DB_DeleteFrom}
\alias{DB_DeleteFrom}
\title{Delete Observations From a Table}
\usage{
DB_DeleteFrom(table)
}
\arguments{
\item{table}{Data frame or table that will be emptied}
}
\value{
A message indicating if the table was successfully deleted.
}
\description{
This function deletes all observations from a table in a PostgreSQL database.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
DB_DeleteFrom(table)
\dontshow{\}) # examplesIf}
}
