% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_fun.R
\name{DB_MetaUpdate}
\alias{DB_MetaUpdate}
\title{Upload Metadata to the Database}
\usage{
DB_MetaUpdate(path)
}
\arguments{
\item{path}{Path to the Excel sheet}
}
\value{
A message if successful
}
\description{
This function reads the manual metadata and uploads
 each sheet as a table in the database.
}
\examples{
\dontrun{
DB_MetaUpdate(path)
}

}
