% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_fun.R
\name{DB_send}
\alias{DB_send}
\title{Send a Table to the Database}
\usage{
DB_send(table, name)
}
\arguments{
\item{table}{Data frame or table that will be uploaded.}

\item{name}{Name of the table.}
}
\value{
A message indicating if the table was successfully uploaded.
}
\description{
This function sends a table to the database.
}
