% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_fun.R
\name{DB_UploadLSS}
\alias{DB_UploadLSS}
\title{Upload LSS Survey Files to the Database}
\usage{
DB_UploadLSS(dir = here::here("Master_2024_18_11"))
}
\arguments{
\item{dir}{Directory where the .lss files are stored.}
}
\value{
A message indicating if the .lss files were successfully uploaded.
}
\description{
This function insert all .lss files in the
 lss_surveys table in the PostgreSQL database.
}
