% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_html}
\alias{extract_html}
\title{Extract Text from HTML}
\usage{
extract_html(input)
}
\arguments{
\item{input}{HTML code}
}
\description{
Some text in LimeSurvey are stored in HTML format.
 This helper function extracts it from the HTML code.
}
\examples{
\dontrun{
extract_html(input = "Html code here")
}
}
