% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtraPlots.R
\name{get_AllExtraPlots}
\alias{get_AllExtraPlots}
\title{Get  Extra Plots for All Report Templates}
\usage{
get_AllExtraPlots(export = FALSE, filter = FALSE)
}
\arguments{
\item{export}{Logical. If TRUE, the function will export the result.}

\item{filter}{Logical. If TRUE, the function returns the extra plots (W) only.}
}
\description{
This is a wrapper function that returns all extra plots based
 on the report_meta_dev data.
}
\examples{
\dontrun{
get_AllExtraPlots(export = FALSE, filter = TRUE)
}
}
