% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtraPlots.R
\name{get_ExtraPlots}
\alias{get_ExtraPlots}
\title{Get Extra Plots}
\usage{
get_ExtraPlots(reporttemplate)
}
\description{
This helper function splits the report_meta_dev data
 and returns the extra plots for a given report template.
}
