% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMaster.R
\name{get_masters}
\alias{get_masters}
\title{Get the templates}
\usage{
get_masters()
}
\description{
Some text in LimeSurvey are stored in HTML format.
 This helper function extracts it from the HTML code.
}
