% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_html}
\alias{is_html}
\title{Test if a Character Vector is actually HTML}
\usage{
is_html(x)
}
\arguments{
\item{x}{A character string}
}
\description{
Some, but not all texts in Lime Survey are formatted as HTML.
 This function uses the `xml2` package to parse the input as HTML.
}
\examples{
extract_html(input = "Html code here")
}
