% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_html}
\alias{is_html}
\title{Test if String is HTML}
\usage{
is_html(x)
}
\arguments{
\item{x}{A character string}
}
\description{
This function tests if the input is HTML. It uses the
 `xml2` package to parse the input as HTML.
}
