% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEW.R
\name{Limer_GetMasterQuesions}
\alias{Limer_GetMasterQuesions}
\title{Get the Quesion from the Master File}
\usage{
Limer_GetMasterQuesions(id, name)
}
\arguments{
\item{id}{Survey ID.}

\item{name}{Survey name (surveyls_title).}
}
\value{
Results from the API.
}
\description{
This function gets the questions (plot, variable, text)
 from the master file.
}
\examples{
\dontrun{
Limer_GetMasterQuesions()
}
}
