% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GET.R
\name{Limer_UploadTemplates}
\alias{Limer_UploadTemplates}
\title{Upload Master Templates as Survey Templates to LimeSurvey}
\usage{
Limer_UploadTemplates()
}
\value{
Results from the API.
}
\description{
This function uploads the master templates as survey templates
to LimeSurvey.
}
\examples{
\dontrun{
Limer_UploadTemplates()
}
}
