% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_delete.R
\name{LS_DeleteSurvey}
\alias{LS_DeleteSurvey}
\title{Delete a Survey from Lime Survey Instance}
\usage{
LS_DeleteSurvey(id)
}
\arguments{
\item{id}{The id of the survey to delete.}
}
\value{
Response from the API.
}
\description{
This function deletes a survey from a Lime Survey instance.
}
\examples{
\dontrun{
LS_DeleteSurvey(id = "116647")
}
}
