% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GET.R
\name{LS_GetMasterQuestions}
\alias{LS_GetMasterQuestions}
\title{Get Questions from a Master File}
\usage{
LS_GetMasterQuestions(id, name)
}
\arguments{
\item{id}{Survey ID.}

\item{name}{Survey name (surveyls_title).}
}
\value{
Results from the API.
}
\description{
This function gets the questions (plot, variable, text)
 from the master survey template.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
LS_GetMasterQuestions(id = '1', name = 'master')
\dontshow{\}) # examplesIf}
}
