% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_GET.R
\name{LS_GetMasterTemplates}
\alias{LS_GetMasterTemplates}
\title{Get Master Templates from Lime Survey}
\usage{
LS_GetMasterTemplates(template = FALSE)
}
\arguments{
\item{template}{Logical. If TRUE, the function will return the template name as well.}
}
\value{
Results from the API.
}
\description{
This function gets the master templates from Lime Survey.
}
\examples{
\dontrun{
LS_GetMasterTemplates()
}
}
