% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limer_Post.R
\name{LS_SendSurveys}
\alias{LS_SendSurveys}
\title{Send Several Survey Templates to Lime Survey}
\usage{
LS_SendSurveys()
}
\value{
Results from the API.
}
\description{
This function is wrapper around LS_SendSurvey to send
 several surveys on the fly.
}
\examples{
\dontrun{
LS_SendSurveys()
}
}
