% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaMasterReport.R
\name{send_testrun}
\alias{send_testrun}
\title{Send results of a test run via Mail}
\usage{
send_testrun(sendto)
}
\arguments{
\item{sendto}{The email address to send the test run to}
}
\description{
This function sends the results of a test run via mail to a
specified email address. It expects a file named "TestRun_YYYY_MM_DD.xlsx".
}
